+++ packages/rocketh-router/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @rocketh/router
 
+## 0.15.0-testing.3
+
+### Patch Changes
+
+- Updated dependencies
+  - rocketh@0.15.0-testing.9
+  - @rocketh/deploy@0.15.0-testing.3
+
 ## 0.15.0-testing.2
 
 ### Patch Changes