+++ packages/rocketh-verifier/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @rocketh/verifier
 
+## 0.15.0-testing.9
+
+### Patch Changes
+
+- Updated dependencies
+  - rocketh@0.15.0-testing.9
+
 ## 0.15.0-testing.8
 
 ### Patch Changes